% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{convert_seconds_to_min_sec_hundredths}
\alias{convert_seconds_to_min_sec_hundredths}
\title{Convert seconds to minute:second.hundredths format}
\usage{
convert_seconds_to_min_sec_hundredths(total_seconds)
}
\arguments{
\item{total_seconds}{A numeric value representing time in seconds}
}
\value{
A string in the format minute:second.hundredths
}
\description{
Converts a given time in seconds to a string formatted as minutes, seconds, and hundredths of a second.
}
\examples{
convert_seconds_to_min_sec_hundredths(980.9)
}
