% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_chat_completion_rtutor}
\alias{create_chat_completion_rtutor}
\title{Create chat completion}
\usage{
create_chat_completion_rtutor(
  model,
  messages = NULL,
  temperature = 1,
  top_p = 1,
  n = 1,
  stream = FALSE,
  stop = NULL,
  max_tokens = NULL,
  presence_penalty = 0,
  frequency_penalty = 0,
  logit_bias = NULL,
  user = NULL,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL,
  seed = NULL
)
}
\arguments{
\item{model}{required; a length one character vector.}

\item{messages}{required; defaults to `NULL`; a list in the following
format: `list(list("role" = "user", "content" = "Hey! How old are you?")`}

\item{temperature}{required; defaults to `1`; a length one numeric vector
with the value between `0` and `2`.}

\item{top_p}{required; defaults to `1`; a length one numeric vector with the
value between `0` and `1`.}

\item{n}{required; defaults to `1`; a length one numeric vector with the
integer value greater than `0`.}

\item{stream}{required; defaults to `FALSE`; a length one logical vector.
**Currently is not implemented.**}

\item{stop}{optional; defaults to `NULL`; a character vector of length
between one and four.}

\item{max_tokens}{required; defaults to `(4096 - prompt tokens)`; a length
one numeric vector with the integer value greater than `0`.}

\item{presence_penalty}{required; defaults to `0`; a length one numeric
vector with a value between `-2` and `2`.}

\item{frequency_penalty}{required; defaults to `0`; a length one numeric
vector with a value between `-2` and `2`.}

\item{logit_bias}{optional; defaults to `NULL`; a named list.}

\item{user}{optional; defaults to `NULL`; a length one character vector.}

\item{openai_api_key}{required; defaults to `Sys.getenv("OPENAI_API_KEY")`
(i.e., the value is retrieved from the `.Renviron` file); a length one
character vector. Specifies OpenAI API key.}

\item{openai_organization}{optional; defaults to `NULL`; a length one
character vector. Specifies OpenAI organization.}

\item{seed}{optional; defaults to `NULL`; a length one
numeric vector with integer values. Attempts deterministic sampling.}
}
\value{
Returns a list, elements of which contain chat completion(s) and
  supplementary information.
}
\description{
Creates a completion for the chat message. See [this
page](https://platform.openai.com/docs/api-reference/chat/create) for
details.
}
\details{
For arguments description please refer to the [official
documentation](https://platform.openai.com/docs/api-reference/chat/create).
}
\examples{
\dontrun{
create_chat_completion(
   model = "gpt-3.5-turbo",
   messages = list(
       list(
           "role" = "system",
           "content" = "You are a helpful assistant."
       ),
       list(
           "role" = "user",
           "content" = "Who won the world series in 2020?"
       ),
       list(
           "role" = "assistant",
           "content" = "The Los Angeles Dodgers won the World Series in 2020."
       ),
       list(
           "role" = "user",
           "content" = "Where was it played?"
       )
   )
)
}
}
